// frontend/src/lib/api/types.gen.ts
// This file is auto-generated by @hey-api/openapi-ts

export type ClientOptions = {
    baseUrl: 'http://localhost:4000' | (string & {});
};

export type GetUsersData = {
    body?: never;
    path?: never;
    query?: never;
    url: '/users/';
};

export type GetUsersResponses = {
    /**
     * Response for status 200
     */
    200: Array<{
        id: number;
        name: string;
        email: string;
    }>;
};

export type GetUsersResponse = GetUsersResponses[keyof GetUsersResponses];

export type PostUsersData = {
    body: {
        name: string;
        email: string;
    };
    path?: never;
    query?: never;
    url: '/users/';
};

export type PostUsersResponses = {
    /**
     * Response for status 200
     */
    200: {
        id: number;
        name: string;
        email: string;
    };
};

export type PostUsersResponse = PostUsersResponses[keyof PostUsersResponses];

export type DeleteUsersByIdData = {
    body?: never;
    path: {
        id: number;
    };
    query?: never;
    url: '/users/{id}';
};

export type GetUsersByIdData = {
    body?: never;
    path: {
        id: number;
    };
    query?: never;
    url: '/users/{id}';
};

export type GetUsersByIdResponses = {
    /**
     * Response for status 200
     */
    200: {
        id: number;
        name: string;
        email: string;
    };
};

export type GetUsersByIdResponse = GetUsersByIdResponses[keyof GetUsersByIdResponses];

export type PatchUsersByIdData = {
    body: {
        name?: string;
        email?: string;
    };
    path: {
        id: number;
    };
    query?: never;
    url: '/users/{id}';
};

export type PatchUsersByIdResponses = {
    /**
     * Response for status 200
     */
    200: {
        id: number;
        name: string;
        email: string;
    };
};

export type PatchUsersByIdResponse = PatchUsersByIdResponses[keyof PatchUsersByIdResponses];